moved file into hat folder, and changed file path to get images
