+++ client/src/components/Hat/hatList.js
@@ -1,0 +1,26 @@
+import banana from '../../../assets/hats/banana.png';
+import bearears from '../../../assets/hats/bearears.png';
+import bubble from '../../../assets/hats/bubble.png';
+import cowboy from '../../../assets/hats/cowboy.png';
+import crown from '../../../assets/hats/crown.png';
+import devil from '../../../assets/hats/devil.png';
+import egg from '../../../assets/hats/egg.png';
+import flowercrown from '../../../assets/hats/flowercrown.png';
+import halo from '../../../assets/hats/halo.png';
+import sprout from '../../../assets/hats/sprout.png';
+
+
+const hatArr = [
+  {id: 1, name: banana},
+  {id: 2, name: bearears},
+  {id: 3, name: bubble},
+  {id: 4, name: cowboy},
+  {id: 5, name: crown},
+  {id: 6, name: devil},
+  {id: 7, name: egg},
+  {id: 8, name: flowercrown},
+  {id: 9, name: halo},
+  {id: 10, name: sprout},
+];
+
+export default hatArr;
